+++ app/concepts/thing/crud.rb
@@ -8,27 +8,6 @@ require "trailblazer/operation/policy"
 require "trailblazer/operation/crud/class_builder"
 
 class Thing < ActiveRecord::Base
-  module SignedIn
-    include Trailblazer::Operation::Module
-
-    contract do
-      property :is_author, virtual: true, default: "0"
-    end
-
-    callback(:before_save) do
-      on_change :add_current_user_as_author!, property: :is_author
-    end
-
-    def add_current_user_as_author!(thing)
-      thing.users << @current_user
-    end
-
-    def setup_params!(params) # TODO: allow passing params to callback.
-      @current_user = params[:current_user]
-    end
-  end
-
-
   class Create < Trailblazer::Operation
     include Resolver
     policy Thing::Policy, :create?
@@ -114,7 +93,6 @@ class Thing < ActiveRecord::Base
     end
 
     class Admin < self
-      include Thing::SignedIn
     end
   end
